add order

+++ src/main/java/com/s2tv/sportshop/model/Order.java
@@ -41,7 +41,7 @@ public class Order {
     boolean isFeedback;
     boolean isPaid;
     Date receivedDate;
-    int orderCode;
+    Long orderCode;
 
     @CreatedDate
     @Field("created_at")
